+++ src/bin.js
@@ -1,24 +1,32 @@
 #!/usr/bin/env node
 import fs from "fs";
+import path from "path";
 import http from "http";
 import { lookup } from "mrmime";
 import { unwatch_all, watch } from "./watch";
 import { name, version } from "../package.json";
 
-let flag = process.argv[2];
-if (["--version", "-v"].includes(flag)) {
+let arg1 = process.argv[2];
+if (["--version", "-v"].includes(arg1)) {
   console.log(`${name}, ${version}`);
   process.exit(0);
 }
-if (["--help", "-h"].includes(flag)) {
+if (["--help", "-h"].includes(arg1)) {
   console.log(`
   Description
     Preview markdown files, update the browser on change.
 
   Usage
-    $ marked-cli
+    $ marked-cli [root]
 `);
   process.exit(0);
+}
+
+// Get base dir from arg1.
+let cwd = process.cwd();
+if (arg1 && fs.existsSync(arg1)) {
+  if (fs.statSync(arg1).isDirectory()) cwd = arg1;
+  else cwd = path.dirname(arg1);
 }
 
 const indexHTML = new URL("index.html", import.meta.url);
@@ -47,17 +55,17 @@ const server = http.createServer((req, res) => {
   // by reading the "path" param of the url.
   // When the event source is closed, we also close the file watch if needed.
   else if (url.pathname === "/@/source") {
+    res.writeHead(200, { "content-type": "text/event-stream" });
+    res.write(`data: 0\n\n`);
     // try resolving markdown file
-    let path = resolve(decodeURIComponent(url.searchParams.get("path") || "/"));
-    if (path && path.endsWith(".md")) {
-      res.once("close", watch(path, send_update.bind(null, res, path)));
-      res.writeHead(200, { "content-type": "text/event-stream" });
-      res.write(`data: 0\n\n`);
+    let file = resolve(decodeURIComponent(url.searchParams.get("path") || "/"));
+    if (file && file.endsWith(".md")) {
+      res.once("close", watch(path.join(cwd, file), send_update.bind(null, res, file)));
       // Send first update.
-      send_update(res, path);
+      send_update(res, file);
     } else {
-      res.statusCode = 404;
-      res.end();
+      res.write(`data: "404 not found"\n\n`);
+      res.once("close", res.end.bind(res));
     }
   }
 
@@ -103,9 +111,9 @@ function create_destroy(server) {
   };
 }
 
-function send_update(res, path) {
+function send_update(res, file) {
   try {
-    let content = fs.readFileSync(path, "utf8");
+    let content = fs.readFileSync(path.join(cwd, file), "utf8");
     res.write(`data: ${JSON.stringify(content)}\n\n`);
   } catch {
     // In case of file being deleted, do force reload.
@@ -115,21 +123,22 @@ function send_update(res, path) {
 
 // This function will be called at least twice for each markdown file.
 // I don't care the performance :)
-function resolve(path) {
-  path = path.replace(/^\//, "");
-  if (path.startsWith("@")) return null;
-  path = path.replace(/\/$/, "");
+function resolve(file) {
+  file = file.replace(/^\//, "");
+  if (file.startsWith("@")) return null;
+  file = file.replace(/\/$/, "");
 
   // exact match
-  if (path) {
-    const exist = fs.existsSync(path);
-    if (exist && fs.statSync(path).isFile()) return path;
+  if (file) {
+    const realpath = path.join(cwd, file);
+    const exist = fs.existsSync(realpath);
+    if (exist && fs.statSync(realpath).isFile()) return file;
     // is folder
-    if (exist) return resolve(path + "/README.md");
+    if (exist) return resolve(file + "/README.md");
   }
 
   // no path, search for README.md in root folder
-  else if (fs.existsSync("README.md")) {
+  else if (fs.existsSync(path.join(cwd, "README.md"))) {
     return "README.md";
   }